# backend/cmd/server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing financial data, chat interactions, and scenario planning",
        "title": "Financial Chat System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v2/fund-flow-rules": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns all fund flow rules with optional filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "List fund flow rules (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by rule type (payment|allocation|transfer)",
                        "name": "ruleType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by target property ID",
                        "name": "targetPropertyId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by target liability ID",
                        "name": "targetLiabilityId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.fundFlowRuleDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Creates a new fund flow rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Create fund flow rule (v2)",
                "parameters": [
                    {
                        "description": "Rule data",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Deletes all fund flow rules for the user",
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Delete all fund flow rules (v2)",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/fund-flow-rules/{id}": {
            "get": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Returns a single fund flow rule by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Get fund flow rule (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Updates an existing fund flow rule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Update fund flow rule (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Rule data",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Deletes a fund flow rule",
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Delete fund flow rule (v2)",
                "operationId": "deleteFundFlowRule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/fund-flow-rules/{id}/stop": {
            "post": {
                "security": [
                    {
                        "SessionID": []
                    },
                    {
                        "AuthToken": []
                    }
                ],
                "description": "Sets an endDate on a rule (soft delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fund Flow Rules V2"
                ],
                "summary": "Stop fund flow rule (v2)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Stop input with endDate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.stopFundFlowRuleDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.fundFlowRuleDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.fundFlowRuleCreateDTO": {
            "type": "object",
            "properties": {
                "amountType": {
                    "description": "Amount",
                    "type": "string"
                },
                "amountValue": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "description": "Priority for multiple rules on same target (lower = higher priority)",
                    "type": "integer"
                },
                "ruleType": {
                    "type": "string"
                },
                "sourceCashAccountId": {
                    "type": "string"
                },
                "sourceCpfAccountId": {
                    "type": "string"
                },
                "sourceIncomeId": {
                    "description": "Source",
                    "type": "string"
                },
                "sourceInvestmentId": {
                    "type": "string"
                },
                "startDate": {
                    "description": "Timing",
                    "type": "string"
                },
                "targetCashAccountId": {
                    "type": "string"
                },
                "targetCpfAccountId": {
                    "description": "Target",
                    "type": "string"
                },
                "targetExpenseId": {
                    "type": "string"
                },
                "targetInvestmentId": {
                    "type": "string"
                },
                "targetLiabilityId": {
                    "type": "string"
                },
                "targetPropertyId": {
                    "type": "string"
                }
            }
        },
        "handlers.fundFlowRuleDTO": {
            "type": "object",
            "properties": {
                "amountType": {
                    "description": "Amount",
                    "type": "string"
                },
                "amountValue": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "Metadata",
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "priority": {
                    "description": "Priority for multiple rules on same target (lower = higher priority)",
                    "type": "integer"
                },
                "ruleType": {
                    "type": "string"
                },
                "sourceCashAccountId": {
                    "type": "string"
                },
                "sourceCpfAccountId": {
                    "type": "string"
                },
                "sourceIncomeId": {
                    "description": "Source",
                    "type": "string"
                },
                "sourceInvestmentId": {
                    "type": "string"
                },
                "startDate": {
                    "description": "Timing",
                    "type": "string"
                },
                "targetCashAccountId": {
                    "type": "string"
                },
                "targetCpfAccountId": {
                    "description": "Target",
                    "type": "string"
                },
                "targetExpenseId": {
                    "type": "string"
                },
                "targetInvestmentId": {
                    "type": "string"
                },
                "targetLiabilityId": {
                    "type": "string"
                },
                "targetPropertyId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthToken": {
            "description": "JWT token for authentication (HMAC-signed with BACKEND_SHARED_SECRET). In dev mode, use GET /api/dev/token?user_id=\u003cuuid\u003e to generate one.",
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}